agent-service/requirements.txt
<<<<<<< Updated upstream
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
=======
annotated-types==0.7.0
anyio==4.13.0
Bottleneck==1.6.0
>>>>>>> Stashed changes
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==47.0.0
<<<<<<< Updated upstream
fastapi==0.136.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.194.0
=======
distro==1.9.0
docstring_parser==0.18.0
google-api-core==2.30.3
>>>>>>> Stashed changes
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.2
idna==3.13
<<<<<<< Updated upstream
=======
iniconfig==2.3.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-core==1.3.2
langchain-google-vertexai==3.2.2
langchain-protocol==0.0.12
langchain-tests==1.1.6
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.11
langgraph-sdk==0.3.13
langsmith==0.7.36
markdown-it-py==4.0.0
mdurl==0.1.2
numexpr==2.14.1
numpy==2.4.4
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.2
pluggy==1.6.0
>>>>>>> Stashed changes
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
<<<<<<< Updated upstream
PyMySQL==1.1.2
pyparsing==3.3.2
python-dotenv==1.2.2
requests==2.33.1
SQLAlchemy==2.0.49
starlette==1.0.0
tqdm==4.67.3
=======
Pygments==2.20.0
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-benchmark==5.2.3
pytest-codspeed==4.4.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
six==1.17.0
sniffio==1.3.1
syrupy==5.1.0
tenacity==9.1.4
>>>>>>> Stashed changes
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
<<<<<<< Updated upstream
uvicorn==0.46.0
=======
uuid_utils==0.14.1
validators==0.35.0
vcrpy==8.1.1
websockets==16.0
wrapt==2.1.2
xxhash==3.7.0
zstandard==0.25.0
>>>>>>> Stashed changes
